pyproject.toml
[project]
name = "agenteval"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain>=1.2.3",
    "langchain-deepseek>=1.0.1",
    "loguru>=0.7.3",
    "python-dotenv>=1.2.1",
]
